// src/interfaces/LoginResponse.ts
// Generated by https://quicktype.io

export interface LoginResponse {
  id: string;
  rol: string;
  username: string;
  email: string;
  token: string;
  refreshToken: string;
  foto: string;
  success: boolean;
  errors: null;
}

// Generated by https://quicktype.io

export interface VerificarToken {
  token: string;
  refreshToken: string;
}

// Generated by https://quicktype.io

export interface LoginRequest {
  email: string;
  password: string;
}

export interface RegisterAccount {
  nombre: string;
  apellidos: string;
  email: string;
  password: string;
  numeroTelefono: string;
  fotoPerfil?: File | null;
}